data/requirements.txt
pydantic==2.6.4
kubernetes==29.0.0
python-dotenv==1.0.1